Overhaul of how warnings work, other minor changes
Warnings used to track only the username and a count of the number of warnings. Now the warning reason, mod who issued the warning, warning date, and a link to the offending post are all in the database. Now each warning is a separate record in the database rather than a single entry with a count. In addition a private wiki page is maintained to visualise the database.

The new mods have also been added to the modlist.

There are other minor tweaks.