Tagged channel update
No longer replies if a tagged channel is not found. Instead it replies with a list of found channels, and messages the user with a list of not found channels.

Also it turns out that reddit's so-called "fancypants" editor adds backslashes everywhere which fuck up my regex. I just filtered them out.